"Add form validation in admin forms for enhanced data integrity"
Updated Brief2SelfReportedAdmin and ChildCBCLAdmin forms in the flourish-child module to ensure enhanced data integrity. Imported json to use in the changeform_view() that checks if certain fields have been correctly filled before form submission. This will prevent the risk of incomplete or inconsistent entries. To achieve this, we added extra_context to the respective admin models.