:recycle: refactor(CRFs): Update child impact default and improve form validation
Changed the default value for the impact on the testing session field to 'no_impact' in child_penn_cnb model and enhanced form submission by using forEach loop instead of for..of loop for better efficiency and readability in the change_form template. Also, made a correction on the brief_2_self_reported_admin model, and updated choices.py with new choices. Introduced a GitHub Action workflow for syncing the forked repository with its parent repo.