:adhesive_bandage: fix(Forms): Fix validation in birth feeding and vaccine form
Added a check for None value for 'total' variable in method of 'BirthFeedingVaccine' form. This fix prevents attempting to create a range from a None value, which would throw a TypeError. This situation might occur when 'infantvaccines_set-TOTAL_FORMS' field is left empty, consequently, improving error handling and robustness of the application.