fix(fix): Remove unnecessary 'cleaned_data' check in pre_flourish_birth_data_form.py
Removed the unnecessary 'cleaned_data' check in the pre_flourish_birth_data_form.py file. The removed lines of code were causing redundancy and extracting the same result from a different data source. The 'initial' data source is sufficient and more reliable in this case. This simplifies the code and reduces the potential for errors.