:sparkles: feature(Models): Add laptop_used field to ChildPennCNB model
Added a new field `laptop_used` to the `ChildPennCNB` model to track which laptop was used for the PennCNB deployment. This will provide more specific data about equipment used for each test, which might be useful for troubleshooting or performance analysis. Additionally, a corresponding `LAPTOP_CHOICES` field was added to `choices.py` to constrained possible laptop entries.