:recycle: refactor(pre_flourish_birth_data_form): Remove unnecessary constructor in pre_flourish_birth_data_form
Removed the constructor method in the pre_flourish_birth_data_form. This constructor was used to get the date of birth from the child consent form and set it as the initial value in the birth data form. However, this logic is no longer necessary because the date of birth is now being handled differently in a more efficient way.