:recycle: refactor(child_clinical_measurements_admin): Refactor a Skin Folds fields section in child_clinical_measurements_admin.py
The fields pertaining to Skin Folds Questions have been restructured, and the phrase 'visit_skin_fold_measure' is moved to 'Child's Waist and Hip Circumference' section. The Main removal of fields under 'Skin Folds Questions' is part of the larger refactoring effort to increase the clarity and simplicity of this section. Also, '1000' and '3000' complete an action of either Remove or Include certain fields under the given numerical section for cleaner data presentation and accessing.