Remove 'child_test_date' field from HIVTesting model
The 'child_test_date' field has been eliminated from the HIVTestingAndResultingMixin model in the 'flourish_child' app. This field was initially capturing the exact testing dates, but it became redundant due to the introduction of the 'child_test_date_estimated' field. This changes allows for better structure and less complexity within the model.