Revise and consolidate age-based form validation
This update revises the form validation functions in child_visit_form.py, consolidating the age-based checks for 7 and 18 years old using the child's age from `caregiver_child_consent_obj`. This eliminates the previous repeated age calculation and improves the readability of the code. This update also streamlines the logic for validating a required model when a participant is not eligible for study participation.