Update signals to handle changes in InfantHIVTesting model
A new signal receiver is added to handle many-to-many field 'test_visit' changes in the InfantHIVTesting model. When selections are added or removed from the 'test_visit' field, the signal triggers an update on the relevant metadata. This is necessary as the changes impact other functions that depend on these field selections.