(refactor): Remove null and blank attributes from HIV test result field
This commit removes the `null` and `blank` attributes from the HIV test result field in the `hiv_testing_and_resulting_mixin.py` file, and replaces them with a `default` attribute. This ensures that every instance of this field in the database has a default value, eliminating potential issues with handling null or blank entries.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>