(refactor): Improve method for prefilling menarche dates in ChildPregTestingForm
This commit refines the method for prefilling menarche dates in the ChildPregTestingForm. The refactoring replaces the previous mechanism of filling in the initial object before instantiation. The update includes a more efficient extraction of relevant fields and attributes from previous instances, as well as optimized rules for setting readonly attributes on forms.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>