(refactor): Improve condition check in child_preg_testing_form
The code in the child_preg_testing_form has been refactored to improve readability and functionality. The condition check for 'menarche_start_est' not only checks for the presence of the key but also ensures its value is not 'NOT_APPLICABLE'. This change enhances the robustness of the initial form checks.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>