(refactor): Update child TB screening model and admin
The child tuberculosis (TB) screening model has been updated to allow null and blank inputs for cases where a child's TB test status is unknown. Additionally, the admin module has been revised to include the FieldsetsModelAdminMixin for PreviousResultsAdminMixin, alongside reordering the sequence of the conditional_fieldlists property function. This improves the flow of information and allows for better data management.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>